extract all task response stuff from user page
